Add kpath centrality

+++ src/graph/betweenness_memory.rs
@@ -2,11 +2,11 @@ use crate::graph::collection_wrappers::{Queue, Stack};
 use crate::graph::sparse_offset_list::SpareOffsetList;
 
 pub struct BetweennessMemory {
-    pub s_stack: Stack<usize>,
-    pub p_list: SpareOffsetList<Vec<usize>>,
+    pub s_stack: Stack<i64>,
+    pub p_list: SpareOffsetList<Vec<i64>>,
     pub sigma: SpareOffsetList<u64>,
     pub d: SpareOffsetList<i64>,
-    pub q: Queue<usize>,
+    pub q: Queue<i64>,
     pub delta: SpareOffsetList<f64>,
 
     // this field is not used, though its purpose is to prevent public instantiation
@@ -17,7 +17,7 @@ impl BetweennessMemory {
     pub fn new() -> BetweennessMemory {
         BetweennessMemory {
             s_stack: Stack::new(),
-            p_list: SpareOffsetList::new(Vec::<usize>::new()),
+            p_list: SpareOffsetList::new(Vec::<i64>::new()),
             sigma: SpareOffsetList::new(0),
             d: SpareOffsetList::new(-1),
             q: Queue::new(),
