+++ src/graph/sparse_offset_list.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 pub struct SpareOffsetList<T: Clone> {
-    map: HashMap<usize, T>,
+    map: HashMap<i64, T>,
     default: T
 }
 
@@ -13,7 +13,7 @@ impl <T: Clone> SpareOffsetList<T> {
         }
     }
 
-    pub fn get(&mut self, index: usize) -> &T {
+    pub fn get(&mut self, index: i64) -> &T {
         if !self.map.contains_key(&index) {
             let value = self.default.clone();
             self.map.insert(index, value);
@@ -22,7 +22,7 @@ impl <T: Clone> SpareOffsetList<T> {
         self.map.get(&index).unwrap()
     }
 
-    pub fn get_mut(&mut self, index: usize) -> &mut T {
+    pub fn get_mut(&mut self, index: i64) -> &mut T {
         if !self.map.contains_key(&index) {
             let value = self.default.clone();
             self.map.insert(index, value);
@@ -31,7 +31,7 @@ impl <T: Clone> SpareOffsetList<T> {
         self.map.get_mut(&index).unwrap()
     }
 
-    pub fn set(&mut self, index: usize, value: T) {
+    pub fn set(&mut self, index: i64, value: T) {
         self.map.insert(index, value);
     }
 }
